Logging can optionally be disabled on test server

+++ tools/src/globals.py
@@ -52,19 +52,27 @@ mandatory_powers = ["red", "blue", "grey"]
     you can set the appropriate importance level of the data
     that are written to your logfiles.
 """
-logger = logging.getLogger()
-logger.setLevel(logging.DEBUG)
-formatter = logging.Formatter(
-    "%(asctime)s :: %(levelname)s :: %(message)s", "%H:%M:%S")
-# logger to file
-if os.path.exists("./logs/game.log"):
-    os.remove("./logs/game.log")
-file_handler = RotatingFileHandler('./logs/game.log', 'a', 1000000, 1)
-file_handler.setLevel(logging.DEBUG)
-file_handler.setFormatter(formatter)
-logger.addHandler(file_handler)
-# logger to console
-if os.environ.get("DONT_LOG_STDOUT") is None:
-    stream_handler = logging.StreamHandler()
-    stream_handler.setLevel(logging.DEBUG)
-    logger.addHandler(stream_handler)
+class DummyLogger():
+    def __getattr__(self, name):
+        return lambda *x: None
+
+
+if os.environ.get("DISABLE_LOGGING") is None:
+    logger = logging.getLogger()
+    logger.setLevel(logging.DEBUG)
+    formatter = logging.Formatter(
+        "%(asctime)s :: %(levelname)s :: %(message)s", "%H:%M:%S")
+    # logger to file
+    if os.path.exists("./logs/game.log"):
+        os.remove("./logs/game.log")
+    file_handler = RotatingFileHandler('./logs/game.log', 'a', 1000000, 1)
+    file_handler.setLevel(logging.DEBUG)
+    file_handler.setFormatter(formatter)
+    logger.addHandler(file_handler)
+    # logger to console
+    if os.environ.get("DONT_LOG_STDOUT") is None:
+        stream_handler = logging.StreamHandler()
+        stream_handler.setLevel(logging.DEBUG)
+        logger.addHandler(stream_handler)
+else:
+    logger = DummyLogger()
